school-infra-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
supabase==2.3.4
python-dotenv==1.0.1
openpyxl==3.1.2
pydantic==2.6.0
httpx==0.26.0
